refactor: RacingService 도입
RacingService를 도입해 Controller의 역할을 줄였음
